+++ test/mysql2_test.rb
@@ -151,7 +151,7 @@ describe "mysql2 connection" do
   end
 
   it "correctly handles ActiveRecord's local timezone setting" do
-    ActiveRecord::Base.default_timezone = :local
+    set_activerecord_timezone(:local)
 
     time = Time.new(2020, 4, 26, 0, 0, 0)