+++ test/sqlite_test.rb
@@ -157,7 +157,7 @@ describe "sqlite3 connection" do
   end if Gem::Version.new(Sequel.version) >= Gem::Version.new("5.33")
 
   it "correctly handles ActiveRecord's local timezone setting" do
-    ActiveRecord::Base.default_timezone = :local
+    set_activerecord_timezone(:local)
 
     time = Time.new(2020, 4, 26, 0, 0, 0)